deepagent/pyproject.toml
[project]
name = "deepagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.6",
    "langchain-deepseek>=1.0.0",
    "langgraph>=1.0.3",
    "tavily-python>=0.7.12",
    "langgraph-checkpoint-postgres>=3.0.1",
    "psycopg[binary,pool]>=3.2.1",
    "langchain-mcp-adapters>=0.1.13",
]
